test_requirements.txt
-e .
colour-runner==0.0.4
coverage==3.7.1
dj_database_url==0.3.0
django==1.7.0
factory-boy==2.4.1
psycopg2==2.5.4
